@@ -86,11 +104,13 @@
                     <artifactId>maven-compiler-plugin</artifactId>
                     <version>${version.maven.compiler.plugin}</version>
                     <configuration>
-                        <forceJavacCompilerUse>true</forceJavacCompilerUse>
-                        <release>11</release>
-                        <source>11</source>
-                        <target>11</target>
+                        <release>17</release>
                     </configuration>
+                </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-surefire-plugin</artifactId>
+                    <version>${version.surefire.plugin}</version>
                 </plugin>
             </plugins>
         </pluginManagement>
