@@ -35,7 +35,7 @@ public class StatisticsData {
     this.recordCount = Optional.ofNullable(breakdown)
         .map(list -> list.stream().mapToInt(StatisticsData::getRecordCount).sum()).orElse(0);
     this.breakdown = Optional.ofNullable(breakdown).filter(list -> !list.isEmpty())
-        .map(ArrayList::new).orElse(null);
+        .map(ArrayList::new).orElse(new ArrayList<>());
   }
 
   /**
@@ -51,7 +51,7 @@ public class StatisticsData {
     this.field = field;
     this.fieldValue = fieldValue;
     this.recordCount = recordCount;
-    this.breakdown = null;
+    this.breakdown = new ArrayList<>();
   }
 
   /**
@@ -82,7 +82,11 @@ public class StatisticsData {
    * of leaf nodes.
    */
   public List<StatisticsData> getBreakdown() {
-    return Collections.unmodifiableList(breakdown);
+    if (breakdown == null) {
+      return Collections.emptyList();
+    } else {
+      return Collections.unmodifiableList(breakdown);
+    }
   }
 
   /**
